Fix problem model test on Windows. Make sure that when we create a path the root of that path is created according to what is expected on the current platform. E.g. C: on Windows.
Summary: Fix problem model test on Windows. Make sure that when we create a path the root of that path is created according to what is expected on the current platform.

Test Plan: Build and test on Windows. See that the test is passing. Build on Linux and verify that test is passing (As before)

Reviewers: #kdevelop, brauch

Reviewed By: #kdevelop, brauch

Subscribers: brauch, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D17131