Fix defines and includes test to run on Windows.
Summary: Fix defines and includes test to run on Windows. Change the projects generator to use the 'new' way of generating project files. That way it's easier to handle the fact that we use different paths on different platforms.

Test Plan: Build on Windows and confirm that test is passing. Build on Linux and confirm that test (test_definesandincludes and test_compilerprovider since they both depend on projectgenerator ) still pass.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D17156