Add scratchpad plugin
Summary:
Adds a scratchpad plugin, which allows you to keep "scratches" of code/text
to experiment or quickly run something without the need for a project.
The plugin adds a new tool-view, which will maintain a list of your scratches
as well as allowing you to compile and run them.
The scratches live in the directory `scratches` in the data directory and
are regular documents so we get all the editing convenience of code-completion
and diagnostics.
Commands used to run them are saved per-scratches and new scratches use the last
used command for that file type/suffix.

FEATURE: 176389

Test Plan:
Add the tool-view on the left tool bar and try creating and using the scratches.
Currently no automated tests.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: aaronpuchert, kfunk, gregormi, brauch, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D16484