Contextbrowser: Ability to show combined problems and decl tooltip
Without this patch, if both a problem and a declaration are available for a
given cursor location, KDevelop would only show the problem tooltip. This makes
the declaration tooltip inaccessible, which can be an annoyance. This patch makes
it show a combined tooltip instead in this case, allowing to access both the
problem(s) and the declaration.

Test Plan: Tested manually.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D18229