Don't crash when project is closed before it was fully opened
When we start loading a project and then directly remove its
root folder, we would never get a projectOpened signal since
we never fully populated the project. But we will see a projectClosing
signal coming in, which would lead to a crash in the
ProjectChangesModel, as we would not have created an item for that
project yet. Prevent this crash by handling the scenario gracefully.