Debugger plugin fixes
Summary:
Rename Qt logging category: common -> debuggercommon

The old name was too generic and undescriptive. This resulted in
the following confusing lines in KDevelop output:
  kdevelop.plugins.common: Change session  QObject(0x0)
  kdevelop.plugins.common: Will reparse arch

MIDebugger: print a specific process error message

Information in the messageboxes is general and sometimes insufficient
to understand why the process could not be started. In such cases
a warning-level error message can be useful. For example:
  Debugger ERRORED QProcess::FailedToStart "execvp: Exec format error"
  Debugger ERRORED QProcess::WriteError "Error writing to process"

GdbDebugger: pass correct string in userCommandOutput()

When Shell script was not empty, both the shell path and the debugger
executable were duplicated in the string sent to userCommandOutput()
because they are prepended to arguments in if(!shell.isEmpty()) branch.

Test Plan:
Run test_gdb, test_gdbprinters, test_lldb and test_lldbformatters.
The number of failed and skipped tests did not change with these fixes.

Reviewers: mwolff

Reviewed By: mwolff

Subscribers: mwolff, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D18325