Don't call clear() on a shared pointer we don't own
As per the documentation this will delete the pointer itself when the
QSharedPointer refcount drops to 0 That can happen while the target
QProcess instance is still running, and will happen occasionally (on
Mac), causing a crash.
The ChooseCMakeInterfaceJob instance will be deleted soon after a failed
connection attempt so there is no need to clear the server pointer here.

BUG: 403644