Make sure we use the same compiler settings as the project is by default
Summary:
Instead of trusting the user to configure it beforehand, make it
possible to check which compiler is being used in the build manager and
uses it to figure out the initial compilation settings.
This includes a QStringList IBSM::findCompiler(item) method to get the
path to the used compiler. I wonder if we should be providing the
sysroot as well.

Test Plan:
Been playing around with the androidqt branch that uses docker
to implement the SDK instead of just expecting the user to install it himself.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mschwarz, aaronpuchert, mwolff, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D11136