cleanup: remove html prefix/suffix feature from navigation widget
This was *only* being used by the project file quickopen feature,
and you only saw it when clicking into the quickopen list, or when
pressing alt. I doubt anyone ever did that. And it was madness
anyways: we decreased the font size in a hacky non-html compliant
way which actually made it quite hard to read for me. So the only
thing we lose is the project name... which is an acceptable
tradeoff. This feature messed up the whole API considerably, removing
it makes me very happy.

We can bring back the project name in the future by introducing a
generic wrapper-widget that supports the
QuickOpenEmbeddedWidgetInterface.