Introduce QuickOpenEmbeddedWidgetCombiner
This mouthful of a class is essentially just a QWidget with a
QVBoxLayout which you can add widgets to. The fun comes from
implementing the QuickOpenEmbeddedWidgetInterface too, by
delegating to the child widgets that got added.

Overall, this allows us to navigate via Alt+Arrow keyboard actions
between combined widgets, as you'd get when you have both, an error
and a declaration at a given cursor declaration.