DVCS Branch Manager with filtering and sorting proposal
Summary:
The porposal is to utilize QSortFilterProxyModel for sorting and filtering.

When working with on a bigger project with multiple branches some means to find a desired branch for switching or creating a new branch may be considered helpful.
So I added a line edit and a QSortFilterProxyModel to do that for me.

Additionally: new branches are currently appended. Though this makes them easy to find, I think it's counter intuitive and a tad ugly. While the sort mechanism of the Proxy does not help initially, it does upon adding branches.

Drawback: the line edit for filtering may be associated with the "New" button by users. This could make the workflow for that less intuitive.

Test Plan:
Fire up Git->Branches on a git project.

Make sure all branches are there.
Type in the line edit to filter. Only branches that match should be included.
Clearing the line edit should result in all elements to become visible again.
Create New branch, that should be added to list in alphabetical order.

Reviewers: apol

Reviewed By: apol

Subscribers: kfunk, apol, kdevelop-devel

Tags: #vdg, #kdevelop

Differential Revision: https://phabricator.kde.org/D20142