astyle: support the system astyle library
Summary:
Find the astyle library installed in the system, and use it instead of
the embedded copy (which is still used if the system library is not
available).

Test Plan:
Builds fine with, and without a system libastyle (the embedded copy is used in the latter case).
`test_astyle` passes in both cases.
I did not try the plugin for real, though.

Reviewers: mwolff, kossebau, apol

Reviewed By: mwolff, apol

Subscribers: lbeltrame, arrowd, cullmann, mwolff, kossebau, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D17760