doc: drop all bits related to installation
In the majority of the cases, users install kdevelop as available in
their distribution, and thus most of this content is not useful; users
that want to compile kdevelop from sources find already instructions in
the development wikis, so no need to duplicate them in a user-taylored
documentation.  Furthermore, the current building instructions are
terribly outdated.