Add working directory to clang parser.
Summary:
This sets the working directory of the clang parser similarly to the way
the build directory is passed to the builder. This allows the parser to
correctly resolve any relative paths in the extra build arguments. This
means relative paths are resolved relative to the build directory (as
they would be resolved by a builder).

Test Plan: In the current test setup it doesn't seem the TestProject allows a TestBuildSystemManager, so it seems like quite a bit of work to create a test for this. (As a result I assume IBuildSystemManager integration isn't well tested in general right now.) If this needs a test, let me know if adding a TestBuildSystemManager is the right way to go.

Reviewers: #kdevelop, aaronpuchert

Reviewed By: aaronpuchert

Subscribers: aaronpuchert, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D22197