[Documentation - QtHelp] Allow navigating with mouse forwards and backwards button
This allows navigating with the mouse forwards and backwards buttons like in
the editor. To get the mouse events an eventFilter is installed on the viewport
of the treeview on the homepage. For the content WebEngineViews a workaround/hack
is needed to access a private child which swallows the events (QTBUG-43602).
Inspiration for that is taken from https://code.qt.io/cgit/qt-creator/qt-creator.git/tree/src/plugins/help/webenginehelpviewer.cpp.