appdata: add <launchable/> entry
Using the pick-session desktop file here, as this keeps the user in
control what session will be started or allows to create a new.
The plain desktop file loading randomly the last used session loaded,
which also could be heavy on resources, might not be wanted at least
by people with many sessions.