clang: Workaround for empty problem ranges at start of document
Sometimes, for some reason we seem to get problem ranges from libclang that
start and end on the first character of the document. As a result, the final
range computed for the ClangProblem will wrongly start from the beginning of the
document and extend until the actual location of the problem. This commit
introduces a workaround which discards the problematic problem ranges.