Fix Infinite Recursion in DUChain
In some projects - I recognized the issue browsing the kbanking part of #kmymoney -
aliased types may point to themselves. I assume an issue with forward-declared types,
but did not investigate this further.
The recursive call of declarationSizeInformation() may then lead to an infinite number
of function calls and consequently to a stack overflow.

The fix ends the recursion if the type is recurring.