Fix ClangUtils::getDefaultArguments when encountering macros
Use ClangUtils::getRawContents instead of manual tokenization. This
is apparently more reliable and gives us the user-written text which
is imo the best we can hope for. I.e. now the user will see e.g.

    foo(int a = INT_MAX)

Which is way more expressive then us displaying the macro value.
We may want to replace more usages of clang_getTokenSpelling with
ClangUtils::getRawContents...

BUG: 369546