Don't get confused when encountering parse errors in default args
When we fail to get a useful string for a default arg, show at least
an explanatory "<parse error>" string in its place.

Sadly, it's not trivial to get the raw string in this case. Doable
maybe, but requires more manual work since the cursor extent won't
include e.g. a mistyped macro name or such.

CCBUG: 369546