This fixes the comment formatting for the Doxygen variants:
* "//!<"
* "/\*!<"
* "/\*\*<"
* "//!"
* "/\*!"
* "///<"

and it keeps the extra "\*" as well as "<" that are part of the comment, which the previous implementation discarded.

NOTE: We are NOT using clang_Cursor_getParsedComment because of https://bugs.llvm.org/show_bug.cgi?id=35333 , that is why we use this for C/C++ comment parsing.

Also, reformat stringhelper tests to improve the readability.