Allow expansion of custom defined variables inside other variables
Currently if a user defined variable is included in another user defined variable
it is replaced by an empty string. Instead we can replace it with the definition
value of the definition. Since this is done in a while loop this continues until
a we hit a fully defined variable, a predefined variable like $PATH or if no
match is found the replacement is still an empty string. This useful for example
if a custom runtime  environment is configured  in some PREFIX. Other variables
like PATH or LD_LIBRARY_PATH can now depend on $PREFIX without needing to type
the value of $PREFIX multiple times.