okteta, projectmanagerview: add explicit title to mainToolBar definition
The toolbar editor seems to not transfer the title from the target
toolbar the plugin toolbars want to merge into by the given name id.
As work-around the title is repeated here, also for consistency with
what the Kate ui.rc definition does and how the toolbars appear in the
toolbar editor.
The title will not be picked up though if toolbars have been already
edited and user copies of the ui.rc files been created, so this only
improves things a bit for untouched installations.