Jump to declaration when definition file doesn't belong to a project
A common scenario for me is Ctrl-clicking on a use of a Qt signal.
When I do that, I do not want to jump to the file generated by moc,
but rather to the header where the signal is declared.

This patch implements this behavior by checking whether we can find
a project for the definition's file. If not, we jump to the
declaration.