Reload project when non-{generated,external} CMake file becomes dirty
Now that we know the list of CMake files included by our project,
we can easily look for non-generated and non-external files and
trigger a reload of the project when one of these changes.

This nicely guards against in-source builds, or build dirs as
child dirs of the source dirs, that generate cmake files. This was
my first approach, which often lead to reload-loops, which I hope
won't be an issue with this approach.