cmake: trim build dir parameters we read from configuration files
We may have written bogus values with trailing newlines in the
past due to a bug that got fixed in the meantime. Still, when the
user tries to open a project opened with a buggy KDevelop version,
we would still load the broken value. To prevent that, trim the
values on the fly when we read them from the configuration file.

Thankfully, we should get away with that as I do not think that
leading/trailing whitespace should be in these configuration
parameters anyhow.