Fix documentation view navigation via forward&back mouse buttons
Before this change forward/back navigation in non-home documentation
pages worked only if KDevelop was built with Qt WebKit, but not with
Qt WebEngine. Unfortunately even after this change it doesn't work with
Qt WebEngine 5.9. It works with Qt 5.14 and probably with some
Qt versions between 5.9 and 5.14. KDevelop might no longer support
Qt 5.9 by the time this feature makes it to a stable release. So I think
making this new feature work with all currently supported Qt versions is
not worth the effort. The non-working new feature doesn't introduce any
regressions for the older Qt versions.

The tryBrowse(Forward|Back) slots prevent crashes when the mouse button
navigation reaches the beginning or the end of history.