Disable extra context menu in Documentation tool view
Right-clicking on a non-home page in Documentation view is supposed to
show context menu entries defined in KDevelop (only "Copy" most of the
time). However currently two context menus are shown when KDevelop is
built with Qt WebEngine: the "Copy" one and on top of it a WebEngine's
native context menu (with actions like "Back", "Forward", "Reload" or
"Open link in a new tab", "Save link"). Dismissing the two context menus
requires two left mouse button clicks, which is surprising and
inconvenient.

Setting context menu policy to Qt::NoContextMenu in addition to the
event filter workaround disables the WebEngine's native context menu.