Show problem descriptions as inline notes at the end of the line
Using KTextEditor's InlineNote interfaces we can show the description of problems at the end of the line. This enables the user to see what's wrong on a glance instead of manual interaction with the problems toolview or moving the mouse to the offending position. Only one problem in each line will be shown in an inline note, but ones with higher severity are shown over ones with lower severity. If the description of a problem doesn't fit in the available with of a line it will get elided. The minimum severity for which an inline note is shown is configurable in language settings where the other options to problem marking are. The default is to show inline notes for warnings and errors.