Do not force update when reloading projects
The parse jobs should be smart enough to figure out what needs
to be reparsed and what not. Most notably, we check whether
the environment matches the one that was used previously. If it
differs, the parse job will continue, otherwise it can stop early.

This patch fixes excessive reparsing times when changing a cmake
file or after manipulating the custom defines/includes for a project.