formattinghelpers: add a failing test
The first comparison in TestAstyle::testFuzzyMatching() tests an
insertion of braces into a single-statement if block. Add a simple right
context to produce a test failure. Any nonempty rightContext (even a
single space) results in a similar test failure, because the last `if`
block of extractFormattedTextFromContext() runs. In this block of code,
text is fuzzy-matched against formattedMergedText and the second closing
brace is discarded, because the first closing brace has been inserted by
the formatter and is absent from unformatted text.

The result of running the test:
FAIL!  : TestAstyle::testFuzzyMatching() Compared values are not the same
   Actual   (extracted): "void b() {// some comment\n    if( a[0] ) {\n        q;\n    }"
   Expected (formatted): "void b() {// some comment\n    if( a[0] ) {\n        q;\n    }\n }\n"

The added test case exposes a major bug in
extractFormattedTextFromContext(): a single inserted brace (with
the matching brace outside of the text fragment and so not inserted)
breaks code. One manifestation of this bug was reported in 2016 as
Bug 365437: the extra brace is inserted when 1TBS Artistic Style is the
current formatter and a user renames a variable.

The bug can also be reproduced by selecting 1TBS Artistic Style for the
C++ language, copying the failing test data (slightly edited to make it
compile) into any C++ source file:

void b() {/*some comment*/
if( "a"[   0] ) void();
 }

void g() {}

Then selecting the test data fragment without the last line (i.e.
without "void g() {}") and triggering the Reformat Source action.